Backend

+++ src/controllers/videoController.js
@@ -1,14 +1,14 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async(req, res) => {
     const videos = await Video.find({});
-    console.log(videos);
     return res.render("home", {pageName: "Home", videos,});
 }
 
 export const watch = async (req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.render("404", {pageName: "Video not found."}); 
     }
@@ -16,20 +16,34 @@ export const watch = async (req, res) => {
 }
 
 export const getEdit = async(req, res) => {
-    const {id} = req.params;
-    const video = await Video.findById(id); 
+    const { id } =req.params;
+    const {
+        user: {_id},
+    } = req.session;
+    const video = await Video.findById(id);
     if(!video) {
-        return res.render("404", {pageName: "Video not found."}); 
+        return res.status(404).render("404", {pageName: "Video not fuound."});
     }
-    return res.render("edit", {pageName: `Edit ${video.title}`, video});
+    if(String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
+    }
+    return res.render("edit", {pageName: `Edit ${video.title}`, video, });
 }
 
 export const postEdit = async(req, res) => {
+    const {
+        user: {_id},
+    } = req.session;
     const {id} = req.params;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({_id:id}); 
+    const myVideo = await Video.findById(id);
+
     if(!video) {
         return res.status(404).res.render("404", {pageName: "Video not found."}); 
+    }
+     if(String(myVideo.owner) !== _id) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -44,25 +58,41 @@ export const getUpload = (req, res) => {
 }
 
 export const postUpload = async(req, res) => {
-    const {title, hashtags, description} = req.body;
+    const {user: {_id}} = req.session;
+    const file = req.file;
+    const {title, hashtags, description } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl: file.path,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
         
     } catch (error) {
-        return res.status(400).res.render("upload", {
-            pageName: `Upload Video`,
-            errorMessage: error._message,
-        });
+        return res.status(404).render("upload",
+         {pageName: "Upload Video",
+         errorMessage: "error!"});
     }
 }
 
 export const deleteVideo = async(req,res) => {
     const { id } = req.params;
+    const {
+        user: {_id},
+    } = req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.status(404).res.render("404", {pageName: "Video not found."}); 
+    }
+    if(String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
@@ -79,3 +109,4 @@ export const search = async(req,res) => {
     }
     return res.render("search", {pageName: "Search", videos});
 };
+
